新device_api接口更新

+++ core/src/main/java/com/ifhz/core/base/commons/anthrity/UserConstants.java
@@ -11,15 +11,15 @@ public class UserConstants {
 
     // 用户类型
     public static final int USER_TYPE_NORMAL = 1; //普通用户
-    public static final int USER_TYPE_MANAGER = 6; //负责人
+    public static final long USER_TYPE_MANAGER = 6l; //负责人
 
     // 用户状态
     public static final int USER_STATUS_ENABLE = 1; //启用
     public static final int USER_STATUS_DISABLE = 2; //禁用
 
     //用户角色
-    public static final int LW_QUERY = 11;//普通用户角色
-    public static final int DB_QUERY = 9; //负责人角色
-    public static final int TY_QUERY = 8; //管理员角色
-    public static final int CP_QUERY = 10; //管理员角色
+    public static final long LW_QUERY = 11l;//普通用户角色
+    public static final long DB_QUERY = 9l; //负责人角色
+    public static final long TY_QUERY = 8l; //管理员角色
+    public static final long CP_QUERY = 10l; //管理员角色
 }
